requirements.txt
Click==7.0

requests==2.24.0

Flask==1.1.1

gunicorn==19.9.0

itsdangerous==1.1.0

Jinja2==2.10.1

MarkupSafe==1.1.1

Werkzeug==0.15.6
bs4
youtube_dl
telethon
cryptg
pillow
hachoir
mega.py

